move middleware to own package and add handler definitions
