+++ internal/handler/handler.go
@@ -1,0 +1,22 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/shivanshkc/authorizer/internal/utils/errutils"
+	"github.com/shivanshkc/authorizer/internal/utils/httputils"
+)
+
+// Handler encapsulates all REST handlers.
+type Handler struct{}
+
+// NotFound handler can be used to serve any unrecognized routes.
+func (h *Handler) NotFound(w http.ResponseWriter, r *http.Request) {
+	httputils.WriteErr(w, errutils.NotFound())
+}
+
+// Health returns 200 if everything is running fine.
+func (h *Handler) Health(w http.ResponseWriter, r *http.Request) {
+	info := map[string]string{}
+	httputils.Write(w, http.StatusOK, nil, info)
+}
