+++ internal/handler/handler_callback.go
@@ -1,0 +1,8 @@
+package handler
+
+import (
+	"net/http"
+)
+
+// Callback handles the provider's OAuth callback.
+func (h *Handler) Callback(w http.ResponseWriter, r *http.Request) {}
