+++ internal/handler/handler_check.go
@@ -1,0 +1,8 @@
+package handler
+
+import (
+	"net/http"
+)
+
+// Check performs an authentication check on the given request.
+func (h *Handler) Check(w http.ResponseWriter, r *http.Request) {}
