+++ internal/handler/handler_redirect.go
@@ -1,0 +1,8 @@
+package handler
+
+import (
+	"net/http"
+)
+
+// Redirect the caller to the specified provider's authentication page.
+func (h *Handler) Redirect(w http.ResponseWriter, r *http.Request) {}
