+++ internal/middleware/mw.go
@@ -1,4 +1,4 @@
-package http
+package middleware
 
 import (
 	"fmt"
