+++ internal/middleware/mw_access.go
@@ -1,4 +1,4 @@
-package http
+package middleware
 
 import (
 	"log/slog"
