+++ internal/middleware/mw_access_test.go
@@ -1,4 +1,4 @@
-package http
+package middleware
 
 import (
 	"bufio"
