+++ internal/middleware/mw_test.go
@@ -1,4 +1,4 @@
-package http
+package middleware
 
 import (
 	"bytes"
